requirements.txt
mcp
fastapi
uvicorn
pydantic
python-dotenv
anthropic
httpx
pytest
pytest-asyncio
pyyaml
